+++ code/system.ui/react.dev.types/src/types/types.mts
@@ -3,6 +3,7 @@ export * from './types.bus.mjs';
 export * from './types.ctx.mjs';
 export * from './types.state.mjs';
 export * from './types.render.mjs';
+export * from './types.render.props.mjs';
 
 export type MarginInput = number | [number] | [number, number] | [number, number, number, number];
 export type Margin = [number, number, number, number];
