+++ code/system.ui/react.dev.types/src/types/types.render.props.mts
@@ -1,0 +1,54 @@
+import { t } from './common';
+
+type Color = string | number;
+
+/**
+ * Rendering state produced by the props.
+ */
+export type DevRenderProps = {
+  component: DevRenderPropsComponent;
+  host: DevRenderPropsHost;
+  debug: DevRenderPropsDebug;
+};
+
+/**
+ * Main Component ("Subject")
+ */
+export type DevRenderPropsComponent = {
+  renderer?: t.DevRendererRef<any>;
+  size?: DevRenderSize;
+  display?: t.DevPropDisplay;
+  backgroundColor?: Color;
+};
+
+/**
+ * Component Host ("Harness")
+ */
+export type DevRenderPropsHost = {
+  backgroundColor?: Color;
+};
+
+/**
+ * Debug Panel
+ */
+export type DevRenderPropsDebug = {
+  main: {
+    renderers: t.DevRendererRef<any>[];
+  };
+};
+
+/**
+ * Size
+ */
+export type DevRenderSize = DevRenderSizeCenter | DevRenderSizeFill;
+export type DevRenderSizeCenter = {
+  mode: 'center';
+  width?: number;
+  height?: number;
+};
+export type DevRenderSizeFill = {
+  mode: 'fill';
+  x: boolean;
+  y: boolean;
+  margin: t.Margin;
+};
